requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
confluent-kafka==2.5.0
databases==0.9.0
Django==5.0.7
django-allauth==0.63.6
django-celery-results==2.5.1
django-messages==0.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.0.3
kombu==5.3.7
mysqlclient==2.2.4
prompt_toolkit==3.0.47
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
